new(vx-demo): refactor legends to codesandbox example
