+++ packages/vx-demo/src/components/Gallery.tsx
@@ -27,7 +27,7 @@ import Pies from '../docs-v2/examples/vx-shape-pie/Example';
 import Trees from './tiles/Trees';
 import Dendrograms from './tiles/Dendrograms';
 import Voronoi from './tiles/Voronoi';
-import Legends from './tiles/Legends';
+import Legends from '../docs-v2/examples/vx-legend/Example';
 import StatsPlot from './tiles/Statsplot';
 import GeoCustom from '../docs-v2/examples/vx-geo-custom/Example';
 import GeoMercator from '../docs-v2/examples/vx-geo-mercator/Example';
