+++ packages/vx-demo/src/docs-v2/examples/vx-legend/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { render } from 'react-dom';
+
+import Example from './Example';
+import './sandbox-styles.css';
+
+render(<Example />, document.getElementById('root'));
