+++ packages/vx-demo/src/docs-v2/examples/vx-legend/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@vx/demo-legends",
+  "description": "Standalone vx legends demo.",
+  "main": "index.tsx",
+  "dependencies": {
+    "@babel/runtime": "^7.8.4",
+    "@types/react": "^16",
+    "@types/react-dom": "^16",
+    "@vx/glyph": "latest",
+    "@vx/legend": "latest",
+    "@vx/scale": "latest",
+    "d3-format": "^1.4.4",
+    "react": "^16",
+    "react-dom": "^16",
+    "react-scripts-ts": "3.1.0",
+    "typescript": "^3"
+  },
+  "keywords": [
+    "visualization",
+    "d3",
+    "react",
+    "vx",
+    "legend"
+  ]
+}
